# don't try parse empty buf