@carllerche 2018-02-26
Thanks for the report. Are you able to verify that this makes a performance difference?

Looking at the code, I would be surprised if it had any noticeable impact. [`request::decode`](https://github.com/tokio-rs/tokio-minihttp/blob/master/src/request.rs#L60) just calls into [`httparse`](https://github.com/seanmonstar/httparse/blob/master/src/lib.rs#L309-L322). Nothing that I see should have any real cost if the buffer is empty.

That said, I would also be fine merging this as it could be clearer.