@carllerche 2018-03-02
Thanks for the research. I think this can still be merged, but it looks like AppVeyor is hung 👻.

I'll try to unblock it.